+++ oinkbrew/HttpClient.cpp
@@ -1,0 +1,125 @@
+#include "HttpClient.h"
+#include "Helper.h"
+#include "spark_wiring_string.h"
+
+
+/**
+* Constructor.
+*/
+HttpClient::HttpClient()
+{
+}
+
+/**
+* Method to send a header, should only be called from within the class.
+*/
+void HttpClient::sendHeader(const char* aHeaderName, const char* aHeaderValue)
+{
+    client.print(aHeaderName);
+    client.print(": ");
+    client.println(aHeaderValue);
+}
+
+void HttpClient::sendHeader(const char* aHeaderName, const int aHeaderValue)
+{
+    client.print(aHeaderName);
+    client.print(": ");
+    client.println(aHeaderValue);
+}
+
+void HttpClient::sendHeader(const char* aHeaderName)
+{
+    client.println(aHeaderName);
+}
+
+/**
+* Method to send an HTTP Request. Allocate variables in your application code
+* in the aResponse struct and set the headers and the options in the aRequest
+* struct.
+*/
+void HttpClient::request(http_request_t &aRequest, http_response_t &aResponse, http_header_t headers[], const char* aHttpMethod)
+{
+    // If a proper response code isn't received it will be set to -1.
+    aResponse.status = -1;
+
+    // NOTE: The default port tertiary statement is unpredictable if the request structure is not initialised
+    // http_request_t request = {0} or memset(&request, 0, sizeof(http_request_t)) should be used
+    // to ensure all fields are zero
+    bool connected = false;
+    if(aRequest.hostname!=NULL) {
+        connected = client.connect(aRequest.hostname.c_str(), (aRequest.port) ? aRequest.port : 80 );
+    }   else {
+
+    	Helper::serialDebug("Connect to IP: ", false);
+    	Helper::serialDebug(Helper::getIpStr(aRequest.ip).c_str(), false);
+    	Helper::serialDebug(":", false);
+    	Helper::serialDebug(String(aRequest.port).c_str(), false);
+
+        connected = client.connect(aRequest.ip, aRequest.port);
+
+        Helper::serialDebug(" -> ", false);
+        Helper::serialDebug(connected);
+    }
+
+    if (!connected) {
+        client.stop();
+        // If TCP Client can't connect to host, exit here.
+        return;
+    }
+
+    //
+    // Send HTTP Headers
+    //
+
+    // Send initial headers (only HTTP 1.0 is supported for now).
+    client.print(aHttpMethod);
+    client.print(" ");
+    client.print(aRequest.path);
+    client.print(" HTTP/1.0\r\n");
+
+    // Send General and Request Headers.
+    sendHeader("Connection", "close"); // Not supporting keep-alive for now.
+    if(aRequest.hostname!=NULL) {
+        sendHeader("HOST", aRequest.hostname.c_str());
+    }
+
+    //Send Entity Headers
+    // TODO: Check the standard, currently sending Content-Length : 0 for empty
+    // POST requests, and no content-length for other types.
+    if (aRequest.body != NULL) {
+        sendHeader("Content-Length", (aRequest.body).length());
+    } else if (strcmp(aHttpMethod, HTTP_METHOD_POST) == 0) { //Check to see if its a Post method.
+        sendHeader("Content-Length", 0);
+    }
+
+    if (headers != NULL)
+    {
+        int i = 0;
+        while (headers[i].header != NULL)
+        {
+            if (headers[i].value != NULL) {
+                sendHeader(headers[i].header, headers[i].value);
+            } else {
+                sendHeader(headers[i].header);
+            }
+            i++;
+        }
+    }
+
+    // Empty line to finish headers
+    client.println();
+    client.flush();
+
+    //
+    // Send HTTP Request Body
+    //
+
+    if (aRequest.body != NULL) {
+        client.println(aRequest.body);
+    }
+
+    // don't care about response
+    // so wait a bit and clean buffer and stop connection
+    delay(100);
+    client.stop();
+}