+++ oinkbrew/Settings.cpp
@@ -30,12 +30,11 @@
 const char OINK_BREW_VERSION[] = "0.1";
 const char OINK_BREW_VERSION_STRING[] = "Version: 0.1";
 
-const char BREWPI_SPARK_REVISION[] = "REV_C";
-
 
 const short REMOTE_LISTENER_PORT = 7872;
 const short LOCAL_LISTENER_PORT = 7873;
 
 
 const unsigned long DURATION_RUN = 1000;
-const unsigned long DURATION_MESSAGE = 120000;
+const unsigned long DURATION_LOG = 15000;
+const unsigned long DURATION_MSG = 120000;